Add success message and update the tests
